+++ src/ZoomHub/Storage/PostgreSQL/Schema/Schema1.hs
@@ -1,52 +1,39 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE OverloadedLabels #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PartialTypeSignatures #-}
 {-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -O0 #-}
+{-# OPTIONS_GHC -Wno-partial-type-signatures #-}
 {-# OPTIONS_GHC -fomit-interface-pragmas #-}
 
 module ZoomHub.Storage.PostgreSQL.Schema.Schema1
-  ( ContentTable1,
+  ( migration,
   )
 where
 
 import Squeal.PostgreSQL
-  ( (:::),
-    (:=>),
-    ColumnConstraint (Def, NoDef),
-    NullityType (NotNull, Null),
-    PGType (PGfloat8, PGint4, PGint8, PGtext, PGtimestamptz),
-    SchemumType (Table),
-    TableConstraint (PrimaryKey, Unique),
+  ( (&),
+    Definition,
+    addColumn,
+    alterTable,
+    dropColumn,
+    nullable,
+    text,
   )
+import Squeal.PostgreSQL.Migration (Migration (..))
+import ZoomHub.Storage.PostgreSQL.Schema.Schema0 (Schemas0)
 
-type ContentTable1 =
-  "content"
-    ::: 'Table
-          ( '[ "pk_content" ::: 'PrimaryKey '["id"],
-               "content_unique_hash_id" ::: 'Unique '["hash_id"],
-               "content_unique_url" ::: 'Unique '["url"]
-             ]
-              :=> '[ "id" ::: 'Def :=> 'NotNull 'PGint8,
-                     "hash_id" ::: 'NoDef :=> 'NotNull 'PGtext,
-                     "type_id" ::: 'Def :=> 'NotNull 'PGint4,
-                     "url" ::: 'NoDef :=> 'NotNull 'PGtext,
-                     "state" ::: 'Def :=> 'NotNull 'PGtext,
-                     "initialized_at" ::: 'Def :=> 'NotNull 'PGtimestamptz,
-                     "active_at" ::: 'Def :=> 'Null 'PGtimestamptz,
-                     "completed_at" ::: 'Def :=> 'Null 'PGtimestamptz,
-                     "title" ::: 'Def :=> 'Null 'PGtext,
-                     "attribution_text" ::: 'Def :=> 'Null 'PGtext,
-                     "attribution_link" ::: 'Def :=> 'Null 'PGtext,
-                     "mime" ::: 'Def :=> 'Null 'PGtext,
-                     "size" ::: 'Def :=> 'Null 'PGint8,
-                     "error" ::: 'Def :=> 'Null 'PGtext,
-                     "progress" ::: 'Def :=> 'NotNull 'PGfloat8,
-                     "abuse_level_id" ::: 'Def :=> 'NotNull 'PGint4,
-                     "num_abuse_reports" ::: 'Def :=> 'NotNull 'PGint8,
-                     "num_views" ::: 'Def :=> 'NotNull 'PGint8,
-                     "version" ::: 'Def :=> 'NotNull 'PGint4,
-                     "submitter_email" ::: 'NoDef :=> 'Null 'PGtext
-                   ]
-          )
+migration :: Migration Definition Schemas0 _
+migration = Migration
+  { name = "2021-02-15-1: Add submitter email",
+    up = setup,
+    down = teardown
+  }
+
+setup :: Definition Schemas0 _
+setup = alterTable #content (addColumn #submitter_email (text & nullable))
+
+teardown :: Definition _ Schemas0
+teardown = alterTable #content (dropColumn #submitter_email)