+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ZoomHub
 
+## 2021-09-09
+
+- Generate verification token for each new submission. This will be used to
+  send an email verification link to the submitter.
+
 ## 2021-02-28
 
 - Require email for each submission: